Add a playground page for the interactive demo, and one for the static demo.
