+++ Example/Playgrounds/Visualizer.playground/Pages/Interactive Example.xcplaygroundpage/Contents.swift
@@ -1,24 +1,6 @@
-//: A UIKit based Playground for presenting user interface
-  
 import PlaygroundSupport
 import Parallaxer
 import RxSwift
-import RxCocoa
-import UIKit
-
-// MARK: Static visualization example.
-
-let visualizer = ParallaxView(frame: CGRect(x: 0, y: 0, width: 500, height: 600))
-
-let parallax = Observable.just(7)
-    .parallax(over: .interval(from: 0, to: 10), visualizer: visualizer)
-    .parallaxRelate(to: .interval(from: -5, to: 15), visualizer: visualizer)
-    .parallaxFocus(on: .interval(from: 0, to: 10), visualizer: visualizer)
-    .parallaxMorph(with: .just(.clampToUnitInterval), visualizer: visualizer)
-    .parallaxMorph(with: .just(.oscillate(numberOfTimes: 3)), visualizer: visualizer)
-
-visualizer
-
 
 // MARK: Interactive visualization example. (Make sure the "live view" is visible in the editor.)
