+++ Example/Playgrounds/Visualizer.playground/contents.xcplayground
@@ -1,4 +1,7 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<playground version='5.0' target-platform='ios' buildActiveScheme='true' executeOnSourceChanges='true'>
-    <timeline fileName='timeline.xctimeline'/>
+<playground version='6.0' target-platform='ios' buildActiveScheme='true'>
+    <pages>
+        <page name='Untitled Page 2'/>
+        <page name='Interactive Visualization'/>
+    </pages>
 </playground>
